Rewrite tests to use new newPost function
This avoids some code duplication
