+++ tests/e2e/page-layouts-modal.spec.js
@@ -4,7 +4,7 @@ import {login} from './tools/lib/login';
 test('checks if the welcome modal on the editor is present and closed when the button is clicked', async ({page, context}) => {
   test.setTimeout(240 * 1000);
   await page.goto('./');
-  await login(page, context);
+  await login(context);
   await page.goto('./wp-admin/post-new.php?post_type=page');
   const locator = page.locator('.components-modal__screen-overlay');
   await expect(locator).toBeVisible();
