+++ tests/e2e/tools/lib/login.js
@@ -1,6 +1,6 @@
 import { nonce } from './nonce';
 
-async function login( page, context ) {
+async function login(context) {
   // Login to admin using request context.
   const response = await context.request.post( './wp-login.php', {
       failOnStatusCode: true,
